Subject: Replaced FlatList with FlashList in UserList component
Body:
Transitioned from using FlatList to FlashList in the UserList component for rendering the user list. This change was motivated by FlashList's superior and optimized performance, especially for large lists. Keep in mind to adjust the estimated item size according to the typical height of the items.

It's essential to note that item and index now require explicit type annotations with the switch to FlashList. We conducted all necessary changes to comply with this requirement in our renderItem function.